solution: refactoring
