+++ src/main/kotlin/io/emeraldpay/dshackle/monitoring/accesslog/Events.kt
@@ -98,148 +98,4 @@ class Events {
             val ts: Instant = Instant.now()
     )
 
-    abstract class BaseBuilder<T>() {
-        companion object {
-            private val remoteIpKeys = listOf(
-                    Metadata.Key.of("x-real-ip", Metadata.ASCII_STRING_MARSHALLER),
-                    Metadata.Key.of("x-forwarded-for", Metadata.ASCII_STRING_MARSHALLER)
-            )
-            private val invalidCharacters = Regex("[\n\t]+")
-        }
-
-        var requestDetails = StreamRequestDetails(
-                UUID.randomUUID(),
-                Instant.now(),
-                Remote(emptyList(), "", "")
-        )
-
-        var chainId: Int = Chain.UNSPECIFIED.id
-        var chain = Chain.UNSPECIFIED
-
-        private fun toInetAddress(ip: String): InetAddress? {
-            val isIp = Character.digit(ip[0], 16) != -1
-            if (!isIp) {
-                return null
-            }
-            return try {
-                InetAddress.getByName(ip)
-            } catch (t: Throwable) {
-                null
-            }
-        }
-
-        private fun findBestIp(ips: List<InetAddress>): InetAddress? {
-            // check if a real remote address is provided, otherwise use any local address
-            return ips.sortedWith(kotlin.Comparator { a, b ->
-                val aLocal = a.isLoopbackAddress || a.isSiteLocalAddress
-                val bLocal = b.isLoopbackAddress || b.isSiteLocalAddress
-                when {
-                    aLocal && bLocal -> 0
-                    aLocal -> 1
-                    else -> -1
-                }
-            }).firstOrNull()
-        }
-
-        private fun clean(s: String): String {
-            return StringUtils.truncate(s, 128)
-                    .replace(invalidCharacters, " ")
-                    .trim()
-        }
-
-        abstract protected fun getT(): T
-
-        fun start(metadata: Metadata, attributes: Attributes): T {
-            val userAgent = metadata.get(Metadata.Key.of("user-agent", Metadata.ASCII_STRING_MARSHALLER))
-                    ?.let(this@BaseBuilder::clean)
-                    ?: ""
-            val ips = ArrayList<InetAddress>()
-            remoteIpKeys.forEach { key ->
-                metadata.get(key)?.let {
-                    it.trim().ifEmpty { null }
-                            ?.let(this@BaseBuilder::toInetAddress)
-                            ?.let(ips::add)
-                }
-            }
-            attributes.get(Grpc.TRANSPORT_ATTR_REMOTE_ADDR)?.let { addr ->
-                if (addr is InetSocketAddress) {
-                    ips.add(addr.address)
-                }
-            }
-            val ip = findBestIp(ips)?.hostAddress ?: ""
-            this.requestDetails = this.requestDetails
-                    .copy(remote = Remote(
-                            ips = ips.map { it.hostAddress },
-                            ip = ip,
-                            userAgent = userAgent
-                    ))
-            return getT()
-        }
-
-        fun withChain(chain: Int): T {
-            this.chainId = chain
-            this.chain = Chain.byId(chainId)
-            return getT()
-        }
-    }
-
-    class SubscribeHeadBuilder() : BaseBuilder<SubscribeHeadBuilder>() {
-        private var index = 0
-
-        override fun getT(): SubscribeHeadBuilder {
-            return this
-        }
-
-        fun onReply(resp: BlockchainOuterClass.ChainHead): SubscribeHead {
-            return SubscribeHead(
-                    chain, UUID.randomUUID(), requestDetails, index++
-            )
-        }
-    }
-
-    class NativeCallBuilder : BaseBuilder<NativeCallBuilder>() {
-
-        val items = ArrayList<NativeCallItemDetails>()
-        val replies = HashMap<Int, NativeCallReplyDetails>()
-
-        override fun getT(): NativeCallBuilder {
-            return this
-        }
-
-        fun onItem(item: BlockchainOuterClass.NativeCallItem): NativeCallBuilder {
-            this.items.add(
-                    NativeCallItemDetails(
-                            item.method,
-                            item.id,
-                            item.payload.size().toLong()
-                    )
-            )
-            return this
-        }
-
-        fun onItemReply(reply: BlockchainOuterClass.NativeCallReplyItem): NativeCallBuilder {
-            this.replies[reply.id] = NativeCallReplyDetails(
-                    reply.id,
-                    reply.succeed,
-                    reply.payload?.size()?.toLong() ?: 0L
-            )
-            return this
-        }
-
-        fun build(): List<NativeCall> {
-            return items.mapIndexed { index, item ->
-                val reply = replies[item.id]
-                NativeCall(
-                        request = requestDetails,
-                        total = items.size,
-                        index = index,
-                        succeed = reply?.succeed ?: false,
-                        blockchain = chain,
-                        nativeCall = item,
-                        payloadSizeBytes = item.payloadSizeBytes,
-                        id = UUID.randomUUID()
-                )
-            }
-        }
-    }
 }
